+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG4YQBZ6HTC3HT7NHFCBSXEA"
+        "01KG6NCWRVEGPNM37SP8Q9K0YG"
     )
     r"""The unique identifier of the data source"""