+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -186,7 +186,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kg4yqbf2jgq3ybhact32m4cj"
+        "conv_01kg6ncw9s3bq8c8b6tk9g2rfg"
     )
     r"""Unique conversation identifier with `conv_` prefix."""