+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG4YQBZ3G95PPQ3J36XXW250"
+        "01KG6NCWRRAEMT3EWNCMCVWZSF"
     )
     r"""The unique identifier of the data source"""