+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG4YQBZ5DQYEXS580KPCBNCN"
+        "01KG6NCWRTXQPEQM4NVF4SCF6C"
     )
     r"""The unique identifier of the data source"""