+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG4YQBZ8Q4ZN1GKKEE1YW1XK"
+        "01KG6NCWRX42828TB5MEVRRVSP"
     )
     r"""The unique identifier of the data source"""