+++ packages/orq-rc/RELEASES.md
@@ -5119,3 +5119,13 @@ Based on:
 - [python v4.3.0-rc.10] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.10] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.10 - packages/orq-rc
+
+## 2026-01-30 05:20:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.11] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.11] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.11 - packages/orq-rc